requirements.txt
praw==7.8.1
prawcore==3.0.2
python-dotenv==1.0.1
